rename trigger to activityHandler
