+++ api/hasura/event_triggers/eventManager.ts
@@ -1,9 +1,9 @@
 import { VercelApiHandler, VercelRequest, VercelResponse } from '@vercel/node';
 
+import activityHandler from '../../../api-lib/event_triggers/activityHandler';
 import checkNomineeDiscord from '../../../api-lib/event_triggers/checkNomineeDiscord';
 import checkNomineeDiscordBot from '../../../api-lib/event_triggers/checkNomineeDiscordBot';
 import checkNomineeTelegram from '../../../api-lib/event_triggers/checkNomineeTelegram';
-import createActivity from '../../../api-lib/event_triggers/createActivity';
 import createCircleCRM from '../../../api-lib/event_triggers/createCircleCRM';
 import createContributionInteractionEvent from '../../../api-lib/event_triggers/createContributionInteractionEvent';
 import createNomineeDiscord from '../../../api-lib/event_triggers/createNomineeDiscord';
@@ -31,9 +31,9 @@ import { verifyHasuraRequestMiddleware } from '../../../api-lib/validate';
 type HandlerDict = { [handlerName: string]: VercelApiHandler };
 
 const HANDLERS: HandlerDict = {
-  usersInsertCreateActivity: createActivity,
-  epochInsertCreateActivity: createActivity,
-  contributionInsertCreateActivity: createActivity,
+  activityHandlerUserInsert: activityHandler,
+  activityHandlerEpochInsert: activityHandler,
+  activityHandlerContributionInsert: activityHandler,
   checkNomineeDiscord,
   checkNomineeDiscordBot,
   checkNomineeTelegram,
