+++ hasura/metadata/databases/default/tables/public_contributions.yaml
@@ -90,7 +90,7 @@ select_permissions:
               _is_null: true
       allow_aggregations: true
 event_triggers:
-  - name: contributionInsertCreateActivity
+  - name: activityHandlerContributionInsert
     definition:
       enable_manual: false
       insert:
@@ -99,7 +99,7 @@ event_triggers:
       interval_sec: 10
       num_retries: 5
       timeout_sec: 60
-    webhook: '{{HASURA_API_BASE_URL}}/event_triggers/eventManager?event=createActivity'
+    webhook: '{{HASURA_API_BASE_URL}}/event_triggers/eventManager?event=activityHandler'
     headers:
       - name: verification_key
         value_from_env: HASURA_EVENT_SECRET
