+++ hasura/metadata/databases/default/tables/public_epoches.yaml
@@ -133,7 +133,7 @@ update_permissions:
               _eq: false
       check: {}
 event_triggers:
-  - name: epochInsertCreateActivity
+  - name: activityHandlerEpochInsert
     definition:
       enable_manual: false
       insert:
@@ -142,7 +142,7 @@ event_triggers:
       interval_sec: 10
       num_retries: 5
       timeout_sec: 60
-    webhook: '{{HASURA_API_BASE_URL}}/event_triggers/eventManager?event=createActivity'
+    webhook: '{{HASURA_API_BASE_URL}}/event_triggers/eventManager?event=activityHandler'
     headers:
       - name: verification_key
         value_from_env: HASURA_EVENT_SECRET
