+++ hasura/metadata/databases/default/tables/public_users.yaml
@@ -216,7 +216,7 @@ event_triggers:
     headers:
       - name: verification_key
         value_from_env: HASURA_EVENT_SECRET
-  - name: usersInsertCreateActivity
+  - name: activityHandlerUserInsert
     definition:
       enable_manual: false
       insert:
@@ -225,7 +225,7 @@ event_triggers:
       interval_sec: 10
       num_retries: 5
       timeout_sec: 60
-    webhook: '{{HASURA_API_BASE_URL}}/event_triggers/eventManager?event=createActivity'
+    webhook: '{{HASURA_API_BASE_URL}}/event_triggers/eventManager?event=activityHandler'
     headers:
       - name: verification_key
         value_from_env: HASURA_EVENT_SECRET
